requirements.txt
anthropic
python-dotenv
prompt_toolkit